fix initial click listener and video autoplay
