+++ packages/engine/src/ecs/classes/EngineEvents.ts
@@ -42,7 +42,8 @@ const EVENTS = {
   START_SUSPENDED_CONTEXTS: 'POSITIONAL_AUDIO_START_SUSPENDED_CONTEXTS' as const,
   SUSPEND_POSITIONAL_AUDIO: 'POSITIONAL_AUDIO_SUSPEND_POSITIONAL_AUDIO' as const,
 
-  BROWSER_NOT_SUPPORTED: 'BROWSER_NOT_SUPPORTED' as const
+  BROWSER_NOT_SUPPORTED: 'BROWSER_NOT_SUPPORTED' as const,
+  SET_USER_HAS_INTERACTED: 'SET_USER_HAS_INTERACTED' as const
 }
 
 /**
